arrange scripts
