+++ scripts/tf_pytest/tests/test.py
@@ -1,5 +1,1 @@
-from tf_pytest import apply, init_destroy
-
-
-def test_public_object(apply):
-    print("ok")
+# pass
