+++ scripts/tf_pytest/tf_pytest/tfstate.py
@@ -1,0 +1,119 @@
+import json
+import logging
+import os
+from abc import ABC, abstractmethod
+
+from utility import exec_cmd as _exec_cmd
+
+# logger
+_logger = logging.getLogger(__name__)
+_logger.setLevel(os.environ.get("LOG_LEVEL", "INFO"))
+
+
+class Node(ABC):
+    @abstractmethod
+    def __init__(self, address, state):
+        self.address = address
+        self.state = state
+
+
+class NodeResources(Node):
+    def __init__(self, address, state):
+        super().__init__(address, state)
+
+    def __getattr__(self, name):
+        _logger.debug("name: {}".format(name))
+        target_addr = (self.address + "." + name).removeprefix(".")
+        _logger.debug("target_addr: {}".format(target_addr))
+
+        _logger.debug("self._state: {}".format(self.state))
+
+        for resource in self.state["resources"]:
+            if resource["address"] == target_addr:
+                return NodeResource(target_addr, resource)
+
+        raise AttributeError("{} is not found".format(target_addr))
+
+
+class NodeDatas(Node):
+    def __init__(self, address, state):
+        super().__init__(address, state)
+
+    def __getattr__(self, name):
+        _logger.debug("name: {}".format(name))
+        target_addr = (self.address + "." + name).removeprefix(".")
+        _logger.debug("target_addr: {}".format(target_addr))
+
+        _logger.debug("self._state: {}".format(self.state))
+
+        new_state = {"resources": [resource for resource in self.state["resources"] if resource["type"] == name]}
+        _logger.debug("new_state: {}".format(new_state))
+        if len(new_state["resources"]) == 0:
+            raise AttributeError("{} is not found".format(target_addr))
+
+        return NodeResources(target_addr, new_state)
+
+
+class NodeResource(Node):
+    def __init__(self, address, state):
+        super().__init__(address, state)
+
+    def __getattr__(self, name):
+        _logger.debug("name: {}".format(name))
+
+        if name in self.state.keys():
+            return self.state[name]
+
+
+class NodeModules(Node):
+    def __init__(self, address, state):
+        super().__init__(address, state)
+
+    def __getattr__(self, name):
+        _logger.debug("name: {}".format(name))
+        target_addr = (self.address + "." + name).removeprefix(".")
+        _logger.debug("target_addr: {}".format(target_addr))
+
+        _logger.debug("self._state: {}".format(self.state))
+
+        for child_module in self.state["child_modules"]:
+            if child_module["address"] == target_addr:
+                return NodeModule(target_addr, child_module)
+
+        raise AttributeError("{} is not found".format(target_addr))
+
+
+class NodeModule(Node):
+    def __init__(self, address, state):
+        super().__init__(address, state)
+
+    def __getattr__(self, name):
+        _logger.debug("name: {}".format(name))
+        target_addr = (self.address + "." + name).removeprefix(".")
+        _logger.debug("target_addr: {}".format(target_addr))
+
+        _logger.debug("self._state: {}".format(self.state))
+
+        if name == "module":
+            return NodeModules(target_addr, {"child_modules": self.state["child_modules"]})
+
+        if name == "data":
+            return NodeDatas(
+                target_addr,
+                {"resources": [resource for resource in self.state["resources"] if resource["mode"] == "data"]},
+            )
+
+        new_state = {"resources": [resource for resource in self.state["resources"] if resource["type"] == name]}
+        _logger.debug("new_state: {}".format(new_state))
+        if len(new_state["resources"]) == 0:
+            raise AttributeError("{} is not found".format(target_addr))
+
+        return NodeResources(target_addr, new_state)
+
+
+class NodeRoot(NodeModule):
+    def __init__(self):
+        cwd = os.environ.get("TF_PYTEST_DIR", "../terraform")
+        state = json.loads(_exec_cmd(["terraform", "show", "-json"], cwd=cwd, print_stderr=True).stdout.decode("utf8"))
+
+        super().__init__("", state["values"]["root_module"])
